Move Issue class to github package (#14)
