+++ src/main/scala/me/rexim/issuestant/Main.scala
@@ -7,6 +7,9 @@ import io.circe.syntax._
 
 import scala.io._
 
+
+import me.rexim.issuestant.github._
+
 case class IssueWithParent(parentNumber: Option[Int], issue: Issue);
 
 object Main {
